Testing in proper place
